feat: complete IS 456 module migration (Phase 4-7)
Migrate remaining modules to codes/is456 namespace:
- detailing.py (591 lines) - beam detailing calculations
- serviceability.py (751 lines) - deflection and crack width checks
- compliance.py (427 lines) - compliance checker orchestration
- ductile.py (127 lines) - ductile detailing requirements

All original modules converted to backward-compatibility stubs.
Private functions and data types explicitly re-exported for tests.
2392 tests passing with zero import changes needed in tests.

Part of TASK-313: Migrate IS 456 specific modules
